Add files via upload
This PID is simpler and easier to change if any part of the code is buggy, because, the main parts are separated into functions. for example, There is only one function to run the motors, another function for computing PIDs. So, if code is buggy, you won't have to copy paste the code for the second motor.